Noticket: fix PMD warnings
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@14094 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/org/ejbca/config/CmpConfiguration.java
@@ -136,7 +136,7 @@ public class CmpConfiguration {
 	}
 	
 	public static int getTCPPortNumber() {
-		return Integer.valueOf(EjbcaConfigurationHolder.getString("cmp.tcp.portno")).intValue();
+		return Integer.valueOf(EjbcaConfigurationHolder.getString("cmp.tcp.portno"));
 	}
 	
 	public static String getTCPLogDir() {
